==> preprocessing_pipeline/3_work_assign.ipynb <==
### Here we assign workplaces to people 
In each polygon we know the number of people. So we generate workplaces there. 

import numpy as np
import pandas as pd 
import matplotlib.pyplot as plt
from shapely.geometry import Point
from tqdm import tqdm
from geopy.distance import great_circle
import random

# SOME AUXILIARY FUNCTIONS
def isSchoolAge(row):
    return row['age'] >= 7 and row['age'] <= 17


def isWorkingAge(row):
    return row['age'] >= 18 and (row['sex'] == 'F' and row['age'] <= 60 or row['sex'] == 'M' and row['age'] <= 55)


def rowPointLocation(row):
    return Point(float(row['latitude']), float(row['longitude']))


def findDistToObj(row, point):
    # finds distances from the selected point to the object represented by DataFrame row
    obj_point = rowPointLocation(row)
    dist = great_circle((point.x, point.y), (obj_point.x, obj_point.y)).km
    return dist

city_name = 'chelyabinsk'

input_dir_workplaces = '../data/initial_' + city_name + '/workplaces_' + city_name + '.txt'
input_dir_people = '../data/initial_' + city_name + '/people_' + city_name + '_splitted_dwellings.txt'
input_dir_households = '../data/initial_' + city_name + '/households.txt'
output_dir = '../data/initial_' + city_name + '/people_' + city_name + '_assigned_workplaces.txt'


people_df = pd.read_csv(input_dir_people, sep='\t')
households_df = pd.read_csv(input_dir_households, sep='\t')
workplaces_df = pd.read_csv(input_dir_workplaces, sep='\t')



hh_points_dic = {}

list_keys = households_df['sp_id'].tolist()
list_values = [Point([row['latitude'], row['longitude']])
               for idx, row in tqdm(households_df.iterrows())]

for key, value in zip(list_keys, list_values):
    hh_points_dic[key] = value

def assignWorkplaces(df_orig, hh_points_dic, workplaces_df_orig) -> pd.DataFrame:

    print("Assigning workplaces started...")
    assign_type = '15km' # 'closest' or any ohter string to apply less than 15km radius workplace
    print("CLOSEST WORKPLACE" if assign_type == 'closest' else "IN DISTANCE <= 15km")
    df = df_orig[df_orig.apply(isWorkingAge, axis=1)]
    
    print('Removed people without work...')
    
    close_work_ids = []
    workers_num = {}
    
    print("Workplaces, total: {}".format(workplaces_df_orig['size'].sum()))
    print("Working people, total: {}".format(df.shape[0]))

    df_add = pd.DataFrame(list(np.zeros(workplaces_df_orig.shape[0])), columns=[
                          'distances'], index=workplaces_df_orig.index)
    
    workplaces_df = workplaces_df_orig.join(df_add)
    print(workplaces_df)
    allWorkplacesFilled = False
    
    print(df.shape)
    for idx in tqdm(range(int(df.shape[0]))):
    # for idx in tqdm(range(100)):
        person_row = df.iloc[idx]
        person_point = hh_points_dic[person_row['sp_hh_id']]
        
        series_distance = workplaces_df.apply(
            findDistToObj, args=(person_point,), axis=1)
        # print(series_distance)
        df_add = pd.DataFrame(series_distance.tolist(), columns=[
                              'distances'], index=workplaces_df.index)
        
        workplaces_df.update(df_add)
        isWorkerAssigned = False
        
        while not isWorkerAssigned:
            # print("Schools: \n", schools_df)
            # print("Work distances: \n", workplaces_df['distances'])
            
            close_workplace = None
            min_index = 0
            ### ASSIGN CLOSEST WORKPLACE
            if (assign_type == 'closest'):
                min_index = workplaces_df['distances'].idxmin()
                close_workplace = workplaces_df.loc[min_index]
                print(close_workplace)
            ### ASSIGN FROM SOME DISTANCE    
            else: 
                workplaceCloseEnough = False 
                counter = 0
                while (not workplaceCloseEnough):
                    random.seed(10)
                    min_index = np.random.choice(workplaces_df['distances'].index, size=1)[0]
                    close_workplace = workplaces_df.loc[min_index]
                    # print(close_workplace['distances'])
                    # print(close_workplace.shape)
                    # print(close_workplace['distances'])
                    counter+=1
                    if counter >= 100:
                        workplaceCloseEnough = True
                        print('Too much tries to choose working place')
                    if (close_workplace['distances'] <= 15): 
                        # print(close_workplace)
                        workplaceCloseEnough = True
            
            ### ASSIGN             
            if not (close_workplace['sp_id'] in workers_num.keys()):  # checking capacity
                workers_num[close_workplace['sp_id']] = 1
            else:
                workers_num[close_workplace['sp_id']
                             ] = workers_num[close_workplace['sp_id']] + 1
                
            if workers_num[close_workplace['sp_id']] == close_workplace['size']:
                # print("Before: {}".format(schools_df.shape[0]))
                # (schools_df.index[min_index])
                # print(min_index)
                
                workplaces_df = workplaces_df.drop(min_index, errors='ignore')
                # print("After: {}".format(schools_df.shape[0]))

                if workplaces_df.shape[0] == 0:
                    allWorkplacesFilled = True
                    
            isWorkerAssigned = True
        
        close_work_ids.append(close_workplace['sp_id'])

        if allWorkplacesFilled:
            print("All workplaces filled to capacity!")
            break
            
        
    df_orig.update(pd.DataFrame(close_work_ids, columns=[
                   'work_id'], index=df.index[:len(close_work_ids)]))    
    df_orig.to_csv(output_dir, sep='\t')
    return df_orig


import time

start = time.time()
df_assigned = assignWorkplaces(people_df, hh_points_dic, workplaces_df)
end = time.time()
print("Time of execution: {} min".format((end - start)/60))

min_idx = workplaces_df['size'].idxmin()
workplaces_df.loc[min_idx]
chosen_idx = np.random.choice(workplaces_df.index, size=1)
workplaces_df = workplaces_df.drop(chosen_idx)
np.random.choice(workplaces_df.index, size=1)[0]

df_assigned['work_id'].unique()

# just some tests
# for i in range(100):
#     person_row = df_assigned.iloc[i]
#     hh_id = person_row['sp_hh_id']
#     work_id = person_row['work_id']
#     if person_row['work_id'] != 'X':
#         work = Point(workplaces_df[workplaces_df['sp_id'] == work_id].latitude, workplaces_df[workplaces_df['sp_id'] == work_id].longitude)  
#         hh =  Point(households_df[households_df['sp_id'] == hh_id].iloc[0].latitude, households_df[households_df['sp_id'] == hh_id].iloc[0].longitude) 
#         dist = great_circle((work.x, work.y), (hh.x, hh.y)).km
#         if (dist > 20):
#             print(person_row, dist)
