==> preprocessing_pipeline/4_generate_schools_data.ipynb <==
### Here we assign schools id to people from population whose age betwen 7 and 17

import numpy as np
import pandas as pd 
import matplotlib.pyplot as plt
from shapely.geometry import Point
from tqdm import tqdm
from geopy.distance import great_circle

city_name = 'chelyabinsk'
input_dir_schools = '../data/initial_' + city_name + '/schools_' + city_name + '.xlsx'
output_dir_schools = '../data/initial_' + city_name + '/schools_' + city_name + '.txt'
schools_df = pd.read_excel(input_dir_schools)
schools_processed = schools_df.drop(columns = ['id сервиса', 'Название'])
schools_processed = schools_df
schools_processed = schools_processed.rename(columns={'Мощность': 'size', 'Широта': 'latitude', 'Долгота': 'longitude'})
schools_processed['sp_id'] = schools_processed.index + 1
columns = schools_processed.columns.tolist()
columns = ['sp_id', 'latitude', 'longitude', 'size']
schools_processed = schools_processed[columns]

print('Total number of schools:', len(schools_processed))
schools_processed.to_csv(output_dir_schools, sep='\t')
